+++ imhotep-server/src/main/native/optimized_ftgs/optimized/src/packed_table.c
@@ -441,7 +441,7 @@ void packed_table_set_cell(packed_table_t *table, int row, int col, long value)
     internal_set_cell(table, row, col, value, packed_vector_index);
 }
 
-long packed_table_get_group(packed_table_t *table, int row)
+int packed_table_get_group(packed_table_t *table, int row)
 {
     int row_size = table->row_size;
